add increase life if enough 3 card item

+++ Mario3/src/objects/materials/BlackBackground.cpp
@@ -15,7 +15,7 @@ void CBlackBackground::Render()
 		x,
 		y,
 		length,
-		CGame::GetInstance()->GetBackBufferHeight()
+		(float)CGame::GetInstance()->GetBackBufferHeight()
 	);
 }
